[Translation] Move ProviderTestCase out of the Tests namespace
